clearing casting warnings
git-svn-id: svn://orca.physics.unc.edu/Orca64@11951 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Custom Hardware/SNO+/XL3/ORXL3Model.m
@@ -2794,7 +2794,7 @@ err:
     [self updateUIFromEcalBundle:hwDic slot:slotNum];
     [self synthesizeFECIntoBundle:&hw_bundle[slotNum] forSlot:slotNum];
     
-    [self setEcal_received:[self ecal_received] | 1UL << slotNum];
+    [self setEcal_received:(uint32_t)([self ecal_received] | 1UL << slotNum)];
     //NSLog(@"ecal received mask: 0x%08x\n", [self ecal_received]);
     if ([self ecal_received] == 0xffffUL) {
         [self ecalToOrcaDocumentsReceived];
